requirements.txt
python-telegram-bot==20.6
openai==0.28
APScheduler==3.10.4